Rename UserService persistence folder
